requirements.txt
pandas
kaleido
plotly.express
